fix: manually apply guardduty changes.
After the last PR failed I realized I had a less than perfect
understanding of how guardduty works and so decided that I wanted to
work through the changes locally.

I've removed the existing module as we don't actually deploy a detector
per region but per delegated account.

I've also added an s3 policy for guardduty to access the publishing
bucket and gave permission to root to modify the kms key that is created
so we don't accidentally brick it.