feat: Migrate guardduty to main module
We now take advantage of the assume roles setup in the org_account to
manage guardduty centrally.

Please note we are not currently managing Members/Invites as I could not
figure out how to get it to work, they are enabled and we will track
them at a future time.

We are also now tracking the detectors in each account and each region
that they are available for.

All this work was applied locally