feat: allow roles to assume for local apply/plan
Previously I had hardcoded my personal IAM User into the module in order
to allow me to plan/apply from my local machine.

This change will allow us to pass in an arbitrary number of roles, it
also allows us to easily add or remove roles without having to mess with
the module, so we can open PR's only as long as we require the ability
to do local work.