feat: consolodate Security Hub findings (#208)
Update Security Hub so that consolodated findings are generated
for resources.   This means that if a resource fails multiple controls
that are checking for the same thing, only one finding will be
generated simplifying the triage and resolution of the problem.

This change required an upgrade to the pinned AWS provider version.